rpc 数据类型

+++ lib/cpp/rpc/Transport.cpp
@@ -5,9 +5,9 @@ int rpc::TcpTransport::sendRequestMsg( const RequestMsg& reqMsg) {
 	BinaryProtocol proto;
 	reqMsg.serialize(&proto);
 
-	uint16 len = uint16(proto.getBuffer().size());
+	int32 len = int32(proto.getBuffer().size());
 	struct evbuffer * msgbuff= evbuffer_new();
-	evbuffer_add(msgbuff,(int8*)&len,MsgHeaderMaxSize);
+	evbuffer_add(msgbuff,(byte*)&len,MsgHeaderMaxSize);
 	evbuffer_add(msgbuff,proto.getBuffer().data(), proto.getBuffer().size());
 
 	int ret = bufferevent_write_buffer(bev_, msgbuff);
@@ -26,7 +26,7 @@ int rpc::TcpTransport::recvResponseMsg(struct evbuffer* buff,std::vector<Respons
 		struct evbuffer_iovec v[1];
 		ret=evbuffer_peek(buff, MsgHeaderMaxSize, NULL, v, 1);
 		RpcAssert(ret==1);
-		uint16 msgLen= *(uint16*)(v[0].iov_base);
+		int32 msgLen= *(int32*)(v[0].iov_base);
 
 		if (evbuffer_get_length(buff) < size_t(msgLen+MsgHeaderMaxSize)) {
 			return ret;
@@ -34,10 +34,10 @@ int rpc::TcpTransport::recvResponseMsg(struct evbuffer* buff,std::vector<Respons
 		evbuffer_drain(buff,MsgHeaderMaxSize);
 		ret+=MsgHeaderMaxSize;
 
-		std::unique_ptr<int8[]> buffArray = std::make_unique<int8[]>(msgLen);
+		std::unique_ptr<byte[]> buffArray = std::make_unique<byte[]>(msgLen);
 		evbuffer_remove(buff, buffArray.get(), msgLen);
 		ret+=msgLen;
-		std::vector<int8> buffVec(buffArray.get(), buffArray.get() + msgLen);
+		std::vector<byte> buffVec(buffArray.get(), buffArray.get() + msgLen);
 
 		BinaryProtocol proto;
 		proto.setBuffer(buffVec);
@@ -53,9 +53,9 @@ int rpc::TcpTransport::sendResponseMsg(const ResponseMsg& respMsg) {
 	BinaryProtocol proto;
 	respMsg.serialize(&proto);
 
-	uint16 len = uint16(proto.getBuffer().size());
+	int32 len = int32(proto.getBuffer().size());
 	struct evbuffer * msgbuff= evbuffer_new();
-	evbuffer_add(msgbuff,(int8*)&len,MsgHeaderMaxSize);
+	evbuffer_add(msgbuff,(byte*)&len,MsgHeaderMaxSize);
 	evbuffer_add(msgbuff,proto.getBuffer().data(), proto.getBuffer().size());
 
 	int ret = bufferevent_write_buffer(bev_, msgbuff);
@@ -74,7 +74,7 @@ int rpc::TcpTransport::recvRequestMsg(struct evbuffer* buff,std::vector<RequestM
 		struct evbuffer_iovec v[1];
 		ret=evbuffer_peek(buff, MsgHeaderMaxSize, NULL, v, 1);
 		RpcAssert(ret==1);
-		uint16 msgLen= *(uint16*)(v[0].iov_base);
+		uint16 msgLen= *(int32*)(v[0].iov_base);
 
 		if (evbuffer_get_length(buff) < size_t( msgLen+MsgHeaderMaxSize)) {
 			return ret;
@@ -82,10 +82,10 @@ int rpc::TcpTransport::recvRequestMsg(struct evbuffer* buff,std::vector<RequestM
 		evbuffer_drain(buff,MsgHeaderMaxSize);
 		ret+=MsgHeaderMaxSize;
 
-		std::unique_ptr<int8[]> buffArray = std::make_unique<int8[]>(msgLen);
+		std::unique_ptr<byte[]> buffArray = std::make_unique<byte[]>(msgLen);
 		evbuffer_remove(buff, buffArray.get(), msgLen);
 		ret+=msgLen;
-		std::vector<int8> buffVec(buffArray.get(), buffArray.get() + msgLen);
+		std::vector<byte> buffVec(buffArray.get(), buffArray.get() + msgLen);
 
 		BinaryProtocol proto;
 		proto.setBuffer(buffVec);
